requirements.txt
pip
discord.py>=1.6
colorlog
motor
pymongo
psutil
validators
